+++ frontend/src/components/DetailedProject/ProductBacklogTab.tsx
@@ -1,10 +1,15 @@
 import Box from "@mui/material/Box";
+import Button from "@mui/material/Button";
 import { SelectChangeEvent } from "@mui/material/Select";
 import { useContext, useMemo, useState } from "react";
 
+import OfflineBacklogGenerator from "../../classes/OfflineBacklogGenerator";
+import { EpicsContext } from "../../contexts/EpicsContextProvider";
+import { ProjectMembersContext } from "../../contexts/ProjectMembersContextProvider";
 import { TicketsContext } from "../../contexts/TicketsContextProvider";
 import Project from "../../types/Project";
 import { TICKET_DRAWER_WIDTH } from "../../utils/constants";
+import download from "../../utils/downloadHtmlDoc";
 import getFilteredTickets from "../../utils/getFilteredTickets";
 import getSortedTickets from "../../utils/getSortedTickets";
 import getTicketsByEpics from "../../utils/getTicketsByEpics";
@@ -18,6 +23,7 @@ import TicketSortSelectDropdown, { TicketSortType } from "../TicketSortSelectDro
 import SearchBar from "../common/SearchBar";
 import SecondaryText from "../common/SecondaryText";
 import TypographySprintInformation from "../common/TypographySprintInformation";
+import DownloadIcon from "../icons/DownloadIcon";
 
 interface ProductBacklogTabProps {
   project: Project;
@@ -27,6 +33,8 @@ const ProductBacklogTab = (props: ProductBacklogTabProps): JSX.Element => {
   const { project } = props;
   const { _id: projectId } = project;
   const { tickets } = useContext(TicketsContext);
+  const { usersMap } = useContext(ProjectMembersContext);
+  const { epicsMap } = useContext(EpicsContext);
   const [searchInput, setSearchInput] = useState<string>("");
   const [selectedTicketId, setSelectedTicketId] = useState<string | null>(null);
   const [sortSelection, setSortSelection] = useState<TicketSortType>("priority-dsc");
@@ -111,11 +119,31 @@ const ProductBacklogTab = (props: ProductBacklogTabProps): JSX.Element => {
     </div>
   );
 
+  const getHtmlString = (): string => {
+    const offlineBacklogGenerator = new OfflineBacklogGenerator(
+      displayedTicketsByEpics,
+      project.title,
+      true,
+      usersMap,
+      epicsMap,
+    );
+
+    return offlineBacklogGenerator.generateHtmlDocument();
+  };
+
   return (
     <>
       <Box sx={{ mr: selectedTicketId ? `${TICKET_DRAWER_WIDTH}px` : "" }}>
         {renderTicketNavbar()}
         <Box p={3} mt={5}>
+          {tickets && tickets.length > 0 && (
+            <Button
+              onClick={() => download(`${project.title}-product-backlog.html`, getHtmlString())}
+              startIcon={<DownloadIcon />}
+            >
+              Download for offline (.html)
+            </Button>
+          )}
           {displayedTicketsByEpics &&
             displayedTicketsByEpics.map((displayedTicketsByEpic) => {
               const { epicId, tickets: displayedTickets } = displayedTicketsByEpic;