+++ frontend/src/components/icons/DownloadIcon.tsx
@@ -1,0 +1,3 @@
+import DownloadIcon from "@mui/icons-material/DownloadForOfflineOutlined";
+
+export default DownloadIcon;